requirements.txt
streamlit
python-dotenv
Pillow
PyPDF2
pydantic
openai
agents
asyncio
numpy
requests
python-pptxstreamlit
python-dotenv
Pillow
PyPDF2
pydantic
openai
agents
asyncio
numpy
requests
python-pptx